+++ napari/utils/settings/_appearance.py
