+++ napari/utils/settings/_application.py
