+++ napari/utils/settings/_base.py
