+++ napari/utils/settings/_experimental.py
