+++ napari/utils/settings/_napari_settings.py
