+++ napari/utils/settings/_plugins.py
