+++ napari/utils/settings/_shortcuts.py
